Move shift logic from AsPCF to FoldToPCFTransform
Still needs work - we want a way to differentiate between the two types of shifts - shifts applied to the resulting pcf to make it well-defined and shifts applied so that the AsPCF coboundary transformation is always invertible. The second type is a transformation that we will want to apply also to the source formula, 'a shift in the lattice' (second type) vs a hop along a single axis (first type).